added log out alert

+++ frontend/admin/admin.js
@@ -22,23 +22,36 @@ let allUserData = [];  //to store all user data (all registered users)
 
 // admin logout
 let logout = document.querySelector("#admin-logout");
-logout.addEventListener("click", async(e)=>{
-	
-	await fetch(`${URL}/user/logout`,{
-			method: "POST",
-  			headers: {"Content-type": "application/json;charset=UTF-8",
-					  "authorization":`${localStorage.getItem("token")}`
-					 },
-			body: JSON.stringify({})
-	 	})
-			.then((res)=>res.json())
-			.then((data)=>{
-				alert("Logging out!");
-				localStorage.clear();
-				allUserData = [];
-				window.location.href = "../index.html"
+logout.addEventListener("click", async (e) => {
+
+	Swal.fire({
+		title: 'Are you sure?',
+		text: "You want to Log Out?",
+		icon: 'warning',
+		showCancelButton: true,
+		confirmButtonColor: '#3085d6',
+		cancelButtonColor: '#d33',
+		confirmButtonText: 'Yes'
+	}).then((result) => {
+		if (result.isConfirmed) {
+			fetch(`${URL}/user/logout`, {
+				method: "POST",
+				headers: {
+					"Content-type": "application/json;charset=UTF-8",
+					"authorization": `${localStorage.getItem("token")}`
+				}
 			})
-			.catch((err)=>console.log(err));
+
+			Swal.fire(
+				'see you soon',
+				'',
+				'success'
+			)
+			localStorage.clear();
+			allUserData = [];
+			window.location.href = "../index.html"
+		}
+	})
 })
 
 
@@ -127,9 +140,9 @@ switchMode.addEventListener('change', function () {
 //new loading admin.html
 dashboardFetch();
 
-async function dashboardFetch(){
-	
-	if(!localStorage.getItem("token")){
+async function dashboardFetch() {
+
+	if (!localStorage.getItem("token")) {
 		// alert("Login fisrt to access Admin Dashboard!");
 		Swal.fire(
 			'Login First!',
@@ -137,61 +150,63 @@ async function dashboardFetch(){
 			'warning'
 		)
 	}
-	else{
+	else {
 		newRequestDiv.style.display = "none";
 		allRegistrationDiv.style.display = "none";
 		bookingDiv.style.display = "none";
 		boxDiv.style.display = "grid";
-	
+
 		// total bookings
-		await fetch(`${URL}/book/`,{
+		await fetch(`${URL}/book/`, {
 			method: "GET",
-				headers: {"Content-type": "application/json;charset=UTF-8",
-						"authorization": `${localStorage.getItem("token")}`
-						}
+			headers: {
+				"Content-type": "application/json;charset=UTF-8",
+				"authorization": `${localStorage.getItem("token")}`
+			}
+		})
+			.then((res) => res.json())
+			.then((data) => {
+				let booknumber = document.querySelector("#total-booking")
+				booknumber.textContent = `${data.data.length}`;
+				// console.log(data.data.length);
 			})
-				.then((res)=>res.json())
-				.then((data)=>{
-					let booknumber = document.querySelector("#total-booking")
-					booknumber.textContent = `${data.data.length}`;
-					// console.log(data.data.length);
-				})
-				.catch((err)=>console.log(err));
+			.catch((err) => console.log(err));
 
 
 		// --------------------all users number-----------------------
 
-		await fetch(`${URL}/user/`,{
+		await fetch(`${URL}/user/`, {
 			method: "GET",
-				headers: {"Content-type": "application/json;charset=UTF-8",
-						"authorization": `${localStorage.getItem("token")}`
-						}
-			})
-				.then((res)=>res.json())
-				.then((data)=>{
-					document.querySelector("#total-registration").textContent = `${data.length}`;
-					let pendingNumber = 0, clientNumber = 0, photographerNumber = 0, blockedNumber = 0;
-					data.forEach((ele)=>{
-						if(ele.isBlocked === true){
-							blockedNumber++;
-						}
-						else if(ele.role === "client"){
-							clientNumber++;
-						}
-						else if(ele.role === "photographer" && ele.approved == true){
-							photographerNumber++;
-						}
-						else if(ele.role === "photographer" && ele.approved == false){
-							pendingNumber++;
-						}
-						
-					})
-					document.querySelector("#total-client").textContent = `${clientNumber}`
-					document.querySelector("#total-photographer").textContent = `${photographerNumber}`
-					document.querySelector("#new-pending-request").textContent = `${pendingNumber}`
-					document.querySelector("#total-blocked").textContent = `${blockedNumber}`
+			headers: {
+				"Content-type": "application/json;charset=UTF-8",
+				"authorization": `${localStorage.getItem("token")}`
+			}
+		})
+			.then((res) => res.json())
+			.then((data) => {
+				document.querySelector("#total-registration").textContent = `${data.length}`;
+				let pendingNumber = 0, clientNumber = 0, photographerNumber = 0, blockedNumber = 0;
+				data.forEach((ele) => {
+					if (ele.isBlocked === true) {
+						blockedNumber++;
+					}
+					else if (ele.role === "client") {
+						clientNumber++;
+					}
+					else if (ele.role === "photographer" && ele.approved == true) {
+						photographerNumber++;
+					}
+					else if (ele.role === "photographer" && ele.approved == false) {
+						pendingNumber++;
+					}
+
 				})
-				.catch((err)=>console.log(err));
+				document.querySelector("#total-client").textContent = `${clientNumber}`
+				document.querySelector("#total-photographer").textContent = `${photographerNumber}`
+				document.querySelector("#new-pending-request").textContent = `${pendingNumber}`
+				document.querySelector("#total-blocked").textContent = `${blockedNumber}`
+			})
+			.catch((err) => console.log(err));
 	}
 }
 
@@ -213,7 +228,7 @@ ds.forEach((ele) => {
 		else if (e.target.innerText === "Booking Orders") {
 			fetchAllBooking();
 		}
-		else if(e.target.innerText === "Dashboard"){
+		else if (e.target.innerText === "Dashboard") {
 			dashboardFetch();
 		}
 	})
@@ -222,28 +237,29 @@ ds.forEach((ele) => {
 
 // fetch all registrations / users
 
-async function fetchAllRegistration(){
-	if(!localStorage.getItem("token")){
+async function fetchAllRegistration() {
+	if (!localStorage.getItem("token")) {
 		Swal.fire(
 			'Login First!',
 			'',
 			'warning'
 		)
 	}
-	else{
-		await fetch(`${URL}/user/`,{
+	else {
+		await fetch(`${URL}/user/`, {
 			method: "GET",
-				  headers: {"Content-type": "application/json;charset=UTF-8",
-						  "authorization": `${localStorage.getItem("token")}`
-						}
-			 })
-				.then((res)=>res.json())
-				.then((data)=>{
-					// console.log(data);
-					allUserData = data;
-					showAllRegistration(data);
-				})
-				.catch((err)=>console.log(err));
+			headers: {
+				"Content-type": "application/json;charset=UTF-8",
+				"authorization": `${localStorage.getItem("token")}`
+			}
+		})
+			.then((res) => res.json())
+			.then((data) => {
+				// console.log(data);
+				allUserData = data;
+				showAllRegistration(data);
+			})
+			.catch((err) => console.log(err));
 	}
 }
 // show all registrations / users
@@ -293,24 +309,24 @@ function showAllRegistration(data) {
 		let td3 = document.createElement("td");
 
 		// let span = document.createElement("span");
-		td3.textContent = ele.role ;
-		if(ele.role === "client"){
-			btr.setAttribute("class","status client");
+		td3.textContent = ele.role;
+		if (ele.role === "client") {
+			btr.setAttribute("class", "status client");
 		}
-		else if(ele.role === "photographer"){
-			btr.setAttribute("class","status photographer");
+		else if (ele.role === "photographer") {
+			btr.setAttribute("class", "status photographer");
 		}
-		else if(ele.role === "admin"){
-			btr.setAttribute("class","status admin");
+		else if (ele.role === "admin") {
+			btr.setAttribute("class", "status admin");
 
 		}
 		// td3.append(span);
 		let td4 = document.createElement("td");
-		if(ele.isBlocked ){
+		if (ele.isBlocked) {
 			td4.textContent = "Blocked";
 			td4.style.color = "red";
 		}
-		else{
+		else {
 			td4.textContent = "Active";
 		}
 		let td5 = document.createElement("td");
@@ -322,11 +338,11 @@ function showAllRegistration(data) {
 		removebtn.style.borderRadius = "40px";
 		removebtn.style.padding = "10px 25px"
 		removebtn.style.border = "none";
-		removebtn.setAttribute("class","redbtn")
-		removebtn.addEventListener("click",()=>{
-			blockUser(ele,"All Registration");
+		removebtn.setAttribute("class", "redbtn")
+		removebtn.addEventListener("click", () => {
+			blockUser(ele, "All Registration");
 		})
-		if(ele.isBlocked){
+		if (ele.isBlocked) {
 			removebtn.style.cursor = "not-allowed";
 		}
 		td5.append(removebtn);
@@ -342,37 +358,38 @@ function showAllRegistration(data) {
 
 
 // fetch all clients
-async function fetchAllClients(){
-	if(!localStorage.getItem("token")){
+async function fetchAllClients() {
+	if (!localStorage.getItem("token")) {
 		Swal.fire(
 			'Login First!',
 			'',
 			'warning'
 		)
 	}
-	else{
-		if(allUserData.length > 1){
-			let allclients = allUserData.filter((ele,index)=>{
+	else {
+		if (allUserData.length > 1) {
+			let allclients = allUserData.filter((ele, index) => {
 				return (ele.role == "client" && ele.isBlocked == false);
 			})
 			showClients(allclients);
 		}
-		else{
+		else {
 			let allclients = [];
-			await fetch(`${URL}/user/`,{
-					method: "GET",
-						headers: {"Content-type": "application/json;charset=UTF-8",
-								"authorization": `${localStorage.getItem("token")}`
-								}
-					})
-					.then((res)=>res.json())
-					.then((data)=>{
-						allclients = data.filter((ele,index)=>{
-							return (ele.role == "client" && ele.isBlocked == false);
-						});
-						showClients(allclients);
-					})
-					.catch((err)=>console.log(err));
+			await fetch(`${URL}/user/`, {
+				method: "GET",
+				headers: {
+					"Content-type": "application/json;charset=UTF-8",
+					"authorization": `${localStorage.getItem("token")}`
+				}
+			})
+				.then((res) => res.json())
+				.then((data) => {
+					allclients = data.filter((ele, index) => {
+						return (ele.role == "client" && ele.isBlocked == false);
+					});
+					showClients(allclients);
+				})
+				.catch((err) => console.log(err));
 		}
 
 	}
@@ -431,9 +448,9 @@ function showClients(data) {
 		removebtn.style.borderRadius = "40px";
 		removebtn.style.padding = "10px 25px"
 		removebtn.style.border = "none";
-		
-		removebtn.addEventListener("click",()=>{
-			blockUser(ele,"All Client");
+
+		removebtn.addEventListener("click", () => {
+			blockUser(ele, "All Client");
 		})
 		td4.append(removebtn);
 
@@ -448,43 +465,44 @@ function showClients(data) {
 
 
 // fetch all photographer
-async function fetchAllPhotographers(){
-	if(!localStorage.getItem("token")){
+async function fetchAllPhotographers() {
+	if (!localStorage.getItem("token")) {
 		Swal.fire(
 			'Login First!',
 			'',
 			'warning'
 		)
 	}
-	else{ 
-		if(allUserData.length > 1){
-			let allphotographers = allUserData.filter((ele,index)=>{
-				return (ele.role == "photographer" && ele.approved == true && ele.isBlocked == false) ;
+	else {
+		if (allUserData.length > 1) {
+			let allphotographers = allUserData.filter((ele, index) => {
+				return (ele.role == "photographer" && ele.approved == true && ele.isBlocked == false);
 			})
 			showPhotographers(allphotographers);
 			// console.log(allphotographers)
 		}
-		else{
+		else {
 			let allphotographers = [];
-			await fetch(`${URL}/user/`,{
-					method: "GET",
-						headers: {"Content-type": "application/json;charset=UTF-8",
-								"authorization": `${localStorage.getItem("token")}`
-								}
-					})
-					.then((res)=>res.json())
-					.then((data)=>{
-						allphotographers = data.filter((ele,index)=>{
-							return (ele.role == "photographer" && ele.approved == true && ele.isBlocked == false);
-						});
-						showPhotographers(allphotographers);
-						// console.log(allphotographers)
-					})
-					.catch((err)=>console.log(err));
+			await fetch(`${URL}/user/`, {
+				method: "GET",
+				headers: {
+					"Content-type": "application/json;charset=UTF-8",
+					"authorization": `${localStorage.getItem("token")}`
+				}
+			})
+				.then((res) => res.json())
+				.then((data) => {
+					allphotographers = data.filter((ele, index) => {
+						return (ele.role == "photographer" && ele.approved == true && ele.isBlocked == false);
+					});
+					showPhotographers(allphotographers);
+					// console.log(allphotographers)
+				})
+				.catch((err) => console.log(err));
 		}
 
 	}
-	
+
 }
 //show all photographers
 function showPhotographers(data) {
@@ -521,7 +539,7 @@ function showPhotographers(data) {
 	th6.textContent = "Price per Hour";
 	let th7 = document.createElement("th");
 	th7.textContent = "Action";
-	theadtr.append(th1,th2,th3,th4,th5,th6,th7);
+	theadtr.append(th1, th2, th3, th4, th5, th6, th7);
 
 	thead.append(theadtr);
 
@@ -545,7 +563,7 @@ function showPhotographers(data) {
 		let td5 = document.createElement("td");
 		td5.textContent = ele.expertise;
 		let td6 = document.createElement("td");
-		td6.textContent ="₹ "+ ele.price ;
+		td6.textContent = "₹ " + ele.price;
 		let td7 = document.createElement("td");
 		let removebtn = document.createElement("button")
 		removebtn.textContent = "Block";
@@ -555,13 +573,13 @@ function showPhotographers(data) {
 		removebtn.style.borderRadius = "40px";
 		removebtn.style.padding = "10px 25px"
 		removebtn.style.border = "none";
-		removebtn.setAttribute("class","redbtn")
-		removebtn.addEventListener("click",()=>{
-			blockUser(ele,"All Photographer");
+		removebtn.setAttribute("class", "redbtn")
+		removebtn.addEventListener("click", () => {
+			blockUser(ele, "All Photographer");
 		})
 		td7.append(removebtn);
 
-		btr.append(td1,td2,td3,td4,td5,td6,td7);
+		btr.append(td1, td2, td3, td4, td5, td6, td7);
 
 		tbody.append(btr);
 	})
@@ -573,33 +591,34 @@ function showPhotographers(data) {
 
 
 //fetch pending / new request
-async function fetchNewRequest(){
-	if(!localStorage.getItem("token")){
+async function fetchNewRequest() {
+	if (!localStorage.getItem("token")) {
 		Swal.fire(
 			'Login First!',
 			'',
 			'warning'
 		)
 	}
-	else{
-		await fetch(`${URL}/user/pending`,{
+	else {
+		await fetch(`${URL}/user/pending`, {
 			method: "GET",
-				  headers: {"Content-type": "application/json;charset=UTF-8",
-						  "authorization": `${localStorage.getItem("token")}`
-						}
-			 })
-				.then((res)=>res.json())
-				.then((data)=>{
-					if(data.message == "Unauthorized"){
-						alert("Login First");
-					}
-					else{
-						showNewRequest(data);
-					}
-				})
-				.catch((err)=>console.log(err));
+			headers: {
+				"Content-type": "application/json;charset=UTF-8",
+				"authorization": `${localStorage.getItem("token")}`
+			}
+		})
+			.then((res) => res.json())
+			.then((data) => {
+				if (data.message == "Unauthorized") {
+					alert("Login First");
+				}
+				else {
+					showNewRequest(data);
+				}
+			})
+			.catch((err) => console.log(err));
 	}
-	 
+
 }
 //show pending / new request
 function showNewRequest(data) {
@@ -690,31 +709,32 @@ function showNewRequest(data) {
 
 
 //fetch all bookings
-async function fetchAllBooking(){
-	if(!localStorage.getItem("token")){
+async function fetchAllBooking() {
+	if (!localStorage.getItem("token")) {
 		Swal.fire(
 			'Login First!',
 			'',
 			'warning'
 		)
 	}
-	else{ 
-		await fetch(`${URL}/book/`,{
+	else {
+		await fetch(`${URL}/book/`, {
 			method: "GET",
-				headers: {"Content-type": "application/json;charset=UTF-8",
-						"authorization": `${localStorage.getItem("token")}`
-						}
+			headers: {
+				"Content-type": "application/json;charset=UTF-8",
+				"authorization": `${localStorage.getItem("token")}`
+			}
+		})
+			.then((res) => res.json())
+			.then((data) => {
+				// console.log(data);
+				showBooking(data.data)
 			})
-				.then((res)=>res.json())
-				.then((data)=>{
-					// console.log(data);
-					showBooking(data.data)
-				})
-				.catch((err)=>console.log(err)); 
+			.catch((err) => console.log(err));
 	}
 }
 //show all bookings
-function showBooking(data){
+function showBooking(data) {
 	// console.log(data);
 
 	allRegistrationDiv.style.display = "none";
@@ -763,7 +783,7 @@ function showBooking(data){
 		// td1.setAttribute("class","tdbooking");
 		let td2 = document.createElement("td");
 
-		td2.textContent = ele.photographer.name ;
+		td2.textContent = ele.photographer.name;
 
 		// td2.setAttribute("class","tdbooking");
 		let td3 = document.createElement("td");
@@ -805,90 +825,91 @@ function showBooking(data){
 // ----------approving photographer request
 async function approveRequest(user) {
 	user.approved = true;
-	await fetch(`${URL}/user/applications/${user.email}`, {    
+	await fetch(`${URL}/user/applications/${user.email}`, {
 		method: "PUT",
 		body: JSON.stringify(user),
-    	headers: {
-        	"Content-type": "application/json; charset=UTF-8",
+		headers: {
+			"Content-type": "application/json; charset=UTF-8",
 			"authorization": `${localStorage.getItem("token")}`
-    	}
+		}
 	})
-	.then(response => response.json())
-	.then(json =>{
-		Swal.fire(
-			'Photographer Request Accepted!',
-			'',
-			'success'
-		)
-	})
-	.catch((err)=>console.log(err))
-	.finally(()=>{
-		fetchNewRequest();
-	})
-		
+		.then(response => response.json())
+		.then(json => {
+			Swal.fire(
+				'Photographer Request Accepted!',
+				'',
+				'success'
+			)
+		})
+		.catch((err) => console.log(err))
+		.finally(() => {
+			fetchNewRequest();
+		})
+
 }
 
 // ----------rejecting photographer request
 async function rejectRequest(user) {
 	user.approved = false;
 
-	await fetch(`${URL}/user/applications/${user.email}`, {    
+	await fetch(`${URL}/user/applications/${user.email}`, {
 		method: "PUT",
 		body: JSON.stringify(user),
-    	headers: {
-        	"Content-type": "application/json; charset=UTF-8",
+		headers: {
+			"Content-type": "application/json; charset=UTF-8",
 			"authorization": `${localStorage.getItem("token")}`
-    	}
+		}
 	})
-	.then(response => response.json())
-	.then(json =>{
-		Swal.fire(
-			'Photographer Request Rejected!',
-			'',
-			'error'
-		)
-	})
-	.catch((err)=>console.log(err))
-	.finally(()=>{
-		fetchNewRequest(); })
+		.then(response => response.json())
+		.then(json => {
+			Swal.fire(
+				'Photographer Request Rejected!',
+				'',
+				'error'
+			)
+		})
+		.catch((err) => console.log(err))
+		.finally(() => {
+			fetchNewRequest();
+		})
 
 }
 
 
-async function blockUser(user,msg){
+async function blockUser(user, msg) {
 	// console.log(user)
-	await fetch(`${URL}/user/block/${user._id}`, {    
+	await fetch(`${URL}/user/block/${user._id}`, {
 		method: "POST",
 		body: JSON.stringify(user),
-    	headers: {
-        	"Content-type": "application/json; charset=UTF-8",
+		headers: {
+			"Content-type": "application/json; charset=UTF-8",
 			"authorization": `${localStorage.getItem("token")}`
-    	}
+		}
 	})
-	.then(response => response.json())
-	.then(json => {
-		// alert("User Blocked!");
-		Swal.fire(
-			'User Blocked!',
-			'',
-			'warning'
-		)
-	})
-	.catch((err)=>alert(err.message))
-	
-		if(msg === "All Registration"){
-			fetchAllRegistration();
-		}
-		else if(msg === "All Client"){
-			fetchAllClients()
-		}
-		else if(msg === "All Photographer"){
-			fetchAllPhotographers()
-		}
-	
+		.then(response => response.json())
+		.then(json => {
+			// alert("User Blocked!");
+			Swal.fire(
+				'User Blocked!',
+				'',
+				'warning'
+			)
+		})
+		.catch((err) => alert(err.message))
+
+	if (msg === "All Registration") {
+		fetchAllRegistration();
+	}
+	else if (msg === "All Client") {
+		fetchAllClients()
+	}
+	else if (msg === "All Photographer") {
+		fetchAllPhotographers()
+	}
+
 }
 
 let readDoc = document.querySelector("#btn-read-doc");
-readDoc.addEventListener("click",(e)=>{
+readDoc.addEventListener("click", (e) => {
 	window.location.href = "https://bookmyshoot-backend.onrender.com/api-docs/"
 })
